requirements.txt
pep8
click
prompt_toolkit
pycryptodome
peewee
tabulate
